pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
flask-debugtoolbar = "^0.16.0"
pytest-cov = "^6.1.1"
pytest-mock = "^3.14.0"
httpx = "^0.28.1"
sqlalchemy-utils = "^0.41.2"
factory-boy = "^3.3.3"
faker = "^37.1.0"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
